+++ tests/conversion_test.py
@@ -13,15 +13,23 @@ def create_source(size, tile_size, dtype, pixel_size):
     return source
 
 
+def load_as_zarr(path, x_um, y_um, w_um, h_um):
+    source = TiffSource(path)
+    data = source.asarray_um(x_um, y_um, w_um, h_um)
+    image = np.asarray(data)
+    image = source.render(image)
+    show_image(image)
+
+
 def save_source_as_ome_tiff(output, source, output_params):
     data = source.asarray()
-    #data = source.get_output_dask()    # significantly slower
+    #data = source.asdask(output_params['tile_size'])
     save_image_as_tiff(source, data, output, output_params, ome=True)
 
 
 def save_source_as_zarr(output, source, output_params):
     data = source.asarray()
-    #data = source.get_output_dask()    # significantly slower
+    #data = source.asdask(output_params['tile_size'])
     save_image_as_ome_zarr(source, data, output, output_params)
 
 
@@ -45,6 +53,7 @@ def conversion_test():
     output_params = {'tile_size': [256, 256], 'npyramid_add': 3, 'pyramid_downsample': 2}
 
     source = create_source(size, [256, 256], np.uint8, [(1, 'um')])
+    nchannels = source.get_nchannels()
     save_source_as_ome_tiff(path1, source, output_params)
 
     source1 = TiffSource(path1)
@@ -61,15 +70,16 @@ def conversion_test():
     data = source3.asarray(pixel_size=[10])
     assert data.ndim == 5
     image, _ = source3.get_yxc_image(data)
-    assert image.ndim == 3
+    assert image.ndim == 3 and image.shape[2] == nchannels
     x0, y0, x1, y1 = size[0] // 2, size[1] // 2, size[0], size[1]
-    tile = source.asarray(x0, y0, x1, y1)
-    assert tile.shape == [1, 3, 1, ..., ...]
-    tile = source.asarray(x0, y0, x1, y1, pixel_size=[10])
-    assert tile.shape == [1, 3, 1, ..., ...]
+    w, h = np.divide(size, 2).astype(int)
+    tile = source3.asarray(x0, y0, x1, y1)
+    assert tile.shape == (1, nchannels, 1, h, w)
+    tile = source3.asarray(x0//10, y0//10, x1//10, y1//10, pixel_size=[10])
+    assert tile.shape == (1, nchannels, 1, h//10, w//10)
 
 
-def check_tiff_arrays(input):
+def check_large_tiff_arrays(input):
     source = TiffSource(input, target_pixel_size=[(10, 'um')])
     image = source.render(source.asarray())
     show_image(image)
@@ -85,9 +95,12 @@ if __name__ == '__main__':
     path3 = 'D:/slides/test.ome.tiff'
     output_params = {'tile_size': [256, 256], 'npyramid_add': 3, 'pyramid_downsample': 2}
 
-    conversion_test()
+    #load_as_zarr('D:/slides/EM04573_01/EM04573_01.ome.tif',
+    #             20400, 13000, 500, 500)
 
-    check_tiff_arrays(path)
+    #conversion_test()
+
+    #check_large_tiff_arrays(path)
 
     progress = tqdm(range(2))
     convert_tiff_to_zarr(path, path2, output_params)